컨트롤러 베이스

+++ src/main/java/com/service/searchplay/model/review/SimpleReview.java
@@ -24,7 +24,7 @@ public class SimpleReview {
     private int review_id; // Autoincrement, Primary Key
 
     @NotNull
-    private Long place_id;
+    private Integer place_id;
 
     @NotNull
     private String user_id;
